refactor(bosses): Ported Araxxor and part of RoTS API from v2 to v3
Completed rewrite of araxxor's getRotations method and started on the Rise of the Six version of the
same command

BREAKING CHANGE: Signatures will change from v2 to v3. As example, in v2, you would run
`api.boss.araxxor.getRotation()`. In v3, the same call will be made as
`api.bosses().araxxor().getRotations()`. The same will be true for the entire library.